Fix bug for the last messages
The last implementation was only bringing the first 5 occurrences of
comments stored. We need to make a search that brings the last 5 stored
documents. This could be done through "sort": "desc", but the "sort": "desc"
results were in a reverse order. To workaround that, this commit iterate
over the slice from its tail to the head.